# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RIS Free ERP API",
        "description": "Open-source Enterprise Resource Planning REST API. Finance, Inventory, HR, Sales, Procurement, Manufacturing, Projects, Quality, Assets, Field Service, and LMS.",
        "contact": {
            "email": "support@tptsolutions.co.nz"
        },
        "license": {
            "name": "Apache-2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/agents": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "List agent profiles",
                "operationId": "dac6470b6ad6cf781bf400811ad0c8e1",
                "responses": {
                    "200": {
                        "description": "Paginated list of agents"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Agents"
                ],
                "summary": "Create an agent profile",
                "operationId": "cb73fc823a69c9218f156b62d808133b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "agent_type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Finance Bot"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "agent_type": {
                                        "type": "string",
                                        "enum": [
                                            "local",
                                            "openrouter",
                                            "api",
                                            "human_subcontractor"
                                        ]
                                    },
                                    "provider_config": {
                                        "description": "Model/API configuration",
                                        "type": "object"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "default": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Agent created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/agents/{id}": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "Get agent with assigned skills",
                "operationId": "5444ee6896ff447f274f04bda74b49d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent profile with skill assignments"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Agents"
                ],
                "summary": "Update agent profile",
                "operationId": "0aecf91c904d29e3fa1e7c755ca17124",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Agents"
                ],
                "summary": "Soft-delete agent profile",
                "operationId": "a3ce2a1d66d9534ddda6629cac7d86a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/v1/agents/{id}/skills": {
            "get": {
                "tags": [
                    "Agents"
                ],
                "summary": "List skills assigned to an agent with registry metadata",
                "operationId": "20cdb1c4adffdb2cb4c740dd0cf07864",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Agent skill assignments with full metadata"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/agents/{id}/skills/{slug}": {
            "put": {
                "tags": [
                    "Agents"
                ],
                "summary": "Enable or disable a skill for an agent",
                "operationId": "6d8ba68a15f0cd24e4b470b0f19fc701",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_enabled": {
                                        "type": "boolean"
                                    },
                                    "config_overrides": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assignment updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Laravel Sanctum token. Obtain via POST /api/auth/login.",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}